+++ controllers/mysql_clustering.go
@@ -4,33 +4,21 @@ import (
 	"context"
 	"database/sql"
 	"fmt"
-	"net/http"
-	"net/url"
-	"strconv"
-	"strings"
 	"time"
 
 	"github.com/cybozu-go/moco"
 	"github.com/cybozu-go/moco/accessor"
 	mocov1alpha1 "github.com/cybozu-go/moco/api/v1alpha1"
-	"github.com/cybozu-go/well"
+	ops "github.com/cybozu-go/moco/operators"
 	"github.com/go-logr/logr"
 	_ "github.com/go-sql-driver/mysql"
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/labels"
 	ctrl "sigs.k8s.io/controller-runtime"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 )
-
-// Operator is the interface for operations
-type Operator interface {
-	Name() string
-	Run(ctx context.Context, infra accessor.Infrastructure, cluster *mocov1alpha1.MySQLCluster, status *accessor.MySQLClusterStatus) error
-}
 
 // Operation defines operations to MySQL Cluster
 type Operation struct {
-	Operators      []Operator
+	Operators      []ops.Operator
 	Wait           bool
 	Conditions     []mocov1alpha1.MySQLClusterCondition
 	SyncedReplicas *int
@@ -98,17 +86,17 @@ func decideNextOperation(log logr.Logger, cluster *mocov1alpha1.MySQLCluster, st
 
 	primaryIndex := selectPrimary(status, cluster)
 
-	ops := updatePrimary(cluster, primaryIndex)
-	if len(ops) != 0 {
+	op := updatePrimary(cluster, primaryIndex)
+	if len(op) != 0 {
 		return &Operation{
-			Operators: ops,
+			Operators: op,
 		}, nil
 	}
 
-	ops = restoreEmptyInstance(status, cluster)
-	if len(ops) != 0 {
+	op = restoreEmptyInstance(status, cluster)
+	if len(op) != 0 {
 		return &Operation{
-			Operators: ops,
+			Operators: op,
 		}, nil
 	}
 
@@ -120,10 +108,10 @@ func decideNextOperation(log logr.Logger, cluster *mocov1alpha1.MySQLCluster, st
 		}, nil
 	}
 
-	ops = configureReplication(status, cluster)
-	if len(ops) != 0 {
+	op = configureReplication(status, cluster)
+	if len(op) != 0 {
 		return &Operation{
-			Operators: ops,
+			Operators: op,
 		}, nil
 	}
 
@@ -136,11 +124,11 @@ func decideNextOperation(log logr.Logger, cluster *mocov1alpha1.MySQLCluster, st
 	}
 
 	syncedReplicas := int(cluster.Spec.Replicas) - len(outOfSyncInts)
-	ops = acceptWriteRequest(status, cluster)
-	if len(ops) != 0 {
+	op = acceptWriteRequest(status, cluster)
+	if len(op) != 0 {
 		return &Operation{
 			Conditions:     availableCondition(outOfSyncInts),
-			Operators:      ops,
+			Operators:      op,
 			SyncedReplicas: &syncedReplicas,
 		}, nil
 	}
@@ -322,50 +310,15 @@ func selectPrimary(status *accessor.MySQLClusterStatus, cluster *mocov1alpha1.My
 	return 0
 }
 
-func updatePrimary(cluster *mocov1alpha1.MySQLCluster, newPrimaryIndex int) []Operator {
+func updatePrimary(cluster *mocov1alpha1.MySQLCluster, newPrimaryIndex int) []ops.Operator {
 	currentPrimaryIndex := cluster.Status.CurrentPrimaryIndex
 	if currentPrimaryIndex != nil && *currentPrimaryIndex == newPrimaryIndex {
 		return nil
 	}
 
-	return []Operator{
-		&updatePrimaryOp{
-			newPrimaryIndex: newPrimaryIndex,
-		},
+	return []ops.Operator{
+		ops.UpdatePrimaryOp(newPrimaryIndex),
 	}
-}
-
-type updatePrimaryOp struct {
-	newPrimaryIndex int
-}
-
-func (o *updatePrimaryOp) Name() string {
-	return moco.OperatorUpdatePrimary
-}
-
-func (o *updatePrimaryOp) Run(ctx context.Context, infra accessor.Infrastructure, cluster *mocov1alpha1.MySQLCluster, status *accessor.MySQLClusterStatus) error {
-	db, err := infra.GetDB(ctx, cluster, o.newPrimaryIndex)
-	if err != nil {
-		return err
-	}
-	cluster.Status.CurrentPrimaryIndex = &o.newPrimaryIndex
-	err = infra.GetClient().Status().Update(ctx, cluster)
-	if err != nil {
-		return err
-	}
-
-	_, err = db.Exec("SET GLOBAL rpl_semi_sync_master_enabled=ON,GLOBAL rpl_semi_sync_slave_enabled=OFF")
-	if err != nil {
-		return err
-	}
-
-	expectedRplSemiSyncMasterWaitForSlaveCount := int(cluster.Spec.Replicas / 2)
-	st := status.InstanceStatus[o.newPrimaryIndex]
-	if st.GlobalVariablesStatus.RplSemiSyncMasterWaitForSlaveCount == expectedRplSemiSyncMasterWaitForSlaveCount {
-		return nil
-	}
-	_, err = db.Exec("SET GLOBAL rpl_semi_sync_master_wait_for_slave_count=?", expectedRplSemiSyncMasterWaitForSlaveCount)
-	return err
 }
 
 func isClonable(state sql.NullString) bool {
@@ -384,21 +337,21 @@ func isCloning(state sql.NullString) bool {
 	return state.String == moco.CloneStatusNotStarted || state.String == moco.CloneStatusInProgress
 }
 
-func restoreEmptyInstance(status *accessor.MySQLClusterStatus, cluster *mocov1alpha1.MySQLCluster) []Operator {
+func restoreEmptyInstance(status *accessor.MySQLClusterStatus, cluster *mocov1alpha1.MySQLCluster) []ops.Operator {
 	primaryIndex := *cluster.Status.CurrentPrimaryIndex
 
 	if status.InstanceStatus[primaryIndex].PrimaryStatus.ExecutedGtidSet == "" {
 		return nil
 	}
 
-	ops := make([]Operator, 0)
+	op := make([]ops.Operator, 0)
 
 	primaryHost := moco.GetHost(cluster, primaryIndex)
 	primaryHostWithPort := fmt.Sprintf("%s:%d", primaryHost, moco.MySQLPort)
 
 	for _, s := range status.InstanceStatus {
 		if !s.GlobalVariablesStatus.CloneValidDonorList.Valid || s.GlobalVariablesStatus.CloneValidDonorList.String != primaryHostWithPort {
-			ops = append(ops, setCloneDonorListOp{})
+			op = append(op, ops.SetCloneDonorListOp())
 			break
 		}
 	}
@@ -409,81 +362,11 @@ func restoreEmptyInstance(status *accessor.MySQLClusterStatus, cluster *mocov1al
 		}
 
 		if isClonable(s.CloneStateStatus.State) && s.PrimaryStatus.ExecutedGtidSet == "" {
-			ops = append(ops, cloneOp{
-				replicaIndex: i,
-			})
+			op = append(op, ops.CloneOp(i))
 		}
 	}
 
-	return ops
-}
-
-type setCloneDonorListOp struct{}
-
-func (setCloneDonorListOp) Name() string {
-	return moco.OperatorSetCloneDonorList
-}
-
-func (setCloneDonorListOp) Run(ctx context.Context, infra accessor.Infrastructure, cluster *mocov1alpha1.MySQLCluster, status *accessor.MySQLClusterStatus) error {
-	primaryHost := moco.GetHost(cluster, *cluster.Status.CurrentPrimaryIndex)
-	primaryHostWithPort := fmt.Sprintf("%s:%d", primaryHost, moco.MySQLAdminPort)
-
-	for i := 0; i < int(cluster.Spec.Replicas); i++ {
-		db, err := infra.GetDB(ctx, cluster, i)
-		if err != nil {
-			return err
-		}
-
-		_, err = db.Exec(`SET GLOBAL clone_valid_donor_list = ?`, primaryHostWithPort)
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-type cloneOp struct {
-	replicaIndex int
-}
-
-func (cloneOp) Name() string {
-	return moco.OperatorClone
-}
-
-func (o cloneOp) Run(ctx context.Context, infra accessor.Infrastructure, cluster *mocov1alpha1.MySQLCluster, status *accessor.MySQLClusterStatus) error {
-	primaryHost := moco.GetHost(cluster, *cluster.Status.CurrentPrimaryIndex)
-	replicaHost := moco.GetHost(cluster, o.replicaIndex)
-
-	req, err := http.NewRequestWithContext(
-		ctx,
-		http.MethodGet,
-		fmt.Sprintf("http://%s:%d/clone", replicaHost, moco.AgentPort),
-		nil,
-	)
-	if err != nil {
-		return err
-	}
-
-	queries := url.Values{
-		moco.CloneParamDonorHostName: []string{primaryHost},
-		moco.CloneParamDonorPort:     []string{strconv.Itoa(moco.MySQLAdminPort)},
-		moco.AgentTokenParam:         []string{cluster.Status.AgentToken},
-	}
-	req.URL.RawQuery = queries.Encode()
-
-	cli := &well.HTTPClient{Client: &http.Client{}}
-	resp, err := cli.Do(req)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK {
-		return fmt.Errorf("failed to clone: %s", resp.Status)
-	}
-
-	return nil
+	return op
 }
 
 func waitForClone(status *accessor.MySQLClusterStatus, cluster *mocov1alpha1.MySQLCluster) (bool, []int) {
@@ -505,10 +388,10 @@ func waitForClone(status *accessor.MySQLClusterStatus, cluster *mocov1alpha1.MyS
 	return count > int(cluster.Spec.Replicas/2), outOfSyncIns
 }
 
-func configureReplication(status *accessor.MySQLClusterStatus, cluster *mocov1alpha1.MySQLCluster) []Operator {
+func configureReplication(status *accessor.MySQLClusterStatus, cluster *mocov1alpha1.MySQLCluster) []ops.Operator {
 	primaryHost := moco.GetHost(cluster, *cluster.Status.CurrentPrimaryIndex)
 
-	var operators []Operator
+	var operators []ops.Operator
 	for i, is := range status.InstanceStatus {
 		if i == *cluster.Status.CurrentPrimaryIndex {
 			continue
@@ -519,66 +402,26 @@ func configureReplication(status *accessor.MySQLClusterStatus, cluster *mocov1al
 		}
 
 		if is.ReplicaStatus == nil || is.ReplicaStatus.MasterHost != primaryHost {
-			operators = append(operators, &configureReplicationOp{
-				Index:       i,
-				PrimaryHost: primaryHost,
-			})
+			operators = append(operators, ops.ConfigureReplicationOp(i, primaryHost))
 		}
 	}
 
 	for i, is := range status.InstanceStatus {
 		if i == *cluster.Status.CurrentPrimaryIndex {
 			if is.Role != moco.PrimaryRole {
-				operators = append(operators, &setLabelsOp{})
+				operators = append(operators, ops.SetLabelsOp())
 				break
 			}
 			continue
 		}
 
 		if is.Role != moco.ReplicaRole {
-			operators = append(operators, &setLabelsOp{})
+			operators = append(operators, ops.SetLabelsOp())
 			break
 		}
 	}
 
 	return operators
-}
-
-type configureReplicationOp struct {
-	Index       int
-	PrimaryHost string
-}
-
-func (r configureReplicationOp) Name() string {
-	return moco.OperatorConfigureReplication
-}
-
-func (r configureReplicationOp) Run(ctx context.Context, infra accessor.Infrastructure, cluster *mocov1alpha1.MySQLCluster, status *accessor.MySQLClusterStatus) error {
-	password, err := moco.GetPassword(ctx, cluster, infra.GetClient(), moco.ReplicationPasswordKey)
-	if err != nil {
-		return err
-	}
-
-	db, err := infra.GetDB(ctx, cluster, r.Index)
-	if err != nil {
-		return err
-	}
-
-	_, err = db.Exec(`STOP SLAVE`)
-	if err != nil {
-		return err
-	}
-	_, err = db.Exec(`CHANGE MASTER TO MASTER_HOST = ?, MASTER_PORT = ?, MASTER_USER = ?, MASTER_PASSWORD = ?, MASTER_AUTO_POSITION = 1`,
-		r.PrimaryHost, moco.MySQLPort, moco.ReplicatorUser, password)
-	if err != nil {
-		return err
-	}
-	_, err = db.Exec("SET GLOBAL rpl_semi_sync_master_enabled=OFF,GLOBAL rpl_semi_sync_slave_enabled=ON")
-	if err != nil {
-		return err
-	}
-	_, err = db.Exec(`START SLAVE`)
-	return err
 }
 
 func waitForReplication(status *accessor.MySQLClusterStatus, cluster *mocov1alpha1.MySQLCluster) (bool, []int) {
@@ -615,60 +458,12 @@ func waitForReplication(status *accessor.MySQLClusterStatus, cluster *mocov1alph
 	return count < int(cluster.Spec.Replicas/2), outOfSyncIns
 }
 
-func acceptWriteRequest(status *accessor.MySQLClusterStatus, cluster *mocov1alpha1.MySQLCluster) []Operator {
+func acceptWriteRequest(status *accessor.MySQLClusterStatus, cluster *mocov1alpha1.MySQLCluster) []ops.Operator {
 	primaryIndex := *cluster.Status.CurrentPrimaryIndex
 
 	if !status.InstanceStatus[primaryIndex].GlobalVariablesStatus.ReadOnly {
 		return nil
 	}
-	return []Operator{
-		&turnOffReadOnlyOp{primaryIndex: primaryIndex}}
-}
-
-type setLabelsOp struct{}
-
-func (setLabelsOp) Name() string {
-	return moco.OperatorSetLabels
-}
-
-func (setLabelsOp) Run(ctx context.Context, infra accessor.Infrastructure, cluster *mocov1alpha1.MySQLCluster, status *accessor.MySQLClusterStatus) error {
-	pods := corev1.PodList{}
-	err := infra.GetClient().List(ctx, &pods, &client.ListOptions{
-		Namespace:     cluster.Namespace,
-		LabelSelector: labels.SelectorFromSet(map[string]string{moco.AppNameKey: moco.UniqueName(cluster)}),
-	})
-	if err != nil {
-		return err
-	}
-
-	for _, pod := range pods.Items {
-		if strings.HasSuffix(pod.Name, strconv.Itoa(*cluster.Status.CurrentPrimaryIndex)) {
-			pod.Labels[moco.RoleKey] = moco.PrimaryRole
-		} else {
-			pod.Labels[moco.RoleKey] = moco.ReplicaRole
-		}
-
-		if err := infra.GetClient().Update(ctx, &pod); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-type turnOffReadOnlyOp struct {
-	primaryIndex int
-}
-
-func (o turnOffReadOnlyOp) Name() string {
-	return moco.OperatorTurnOffReadOnly
-}
-
-func (o turnOffReadOnlyOp) Run(ctx context.Context, infra accessor.Infrastructure, cluster *mocov1alpha1.MySQLCluster, status *accessor.MySQLClusterStatus) error {
-	db, err := infra.GetDB(ctx, cluster, o.primaryIndex)
-	if err != nil {
-		return err
-	}
-	_, err = db.Exec("set global read_only=0")
-	return err
+	return []ops.Operator{
+		ops.TurnOffReadOnlyOp(primaryIndex)}
 }
