+++ operators/operator.go
@@ -1,0 +1,14 @@
+package operators
+
+import (
+	"context"
+
+	"github.com/cybozu-go/moco/accessor"
+	mocov1alpha1 "github.com/cybozu-go/moco/api/v1alpha1"
+)
+
+// Operator is the interface for operations
+type Operator interface {
+	Name() string
+	Run(ctx context.Context, infra accessor.Infrastructure, cluster *mocov1alpha1.MySQLCluster, status *accessor.MySQLClusterStatus) error
+}
